+++ src/ore/BUILD.bazel
@@ -62,6 +62,7 @@ rust_library(
         "process",
         "prometheus",
         "proptest",
+        "proptest-derive",
         "rand",
         "region",
         "sentry",
@@ -146,6 +147,7 @@ rust_test(
         "process",
         "prometheus",
         "proptest",
+        "proptest-derive",
         "rand",
         "region",
         "sentry",
