+++ src/ore/Cargo.toml
@@ -49,6 +49,7 @@ prometheus = { version = "0.13.3", default-features = false, optional = true }
 proptest = { version = "1.6.0", default-features = false, features = [
     "std",
 ], optional = true }
+proptest-derive = { version = "0.5.1", optional = true }
 rand = { version = "0.8.5", optional = true }
 smallvec = { version = "1.14.0", optional = true }
 stacker = { version = "0.1.19", optional = true }
@@ -152,6 +153,7 @@ metrics = ["prometheus", "derivative", "assert"]
 id_gen = ["hibitset", "rand", "uuid"]
 capture = ["tracing-capture"]
 assert = ["derivative", "ctor"]
+proptest = ["dep:proptest", "proptest-derive"]
 
 [[test]]
 name = "future"
