+++ src/repr/src/scalar.rs
@@ -20,8 +20,8 @@ use enum_kinds::EnumKind;
 use itertools::Itertools;
 use mz_lowertest::MzReflect;
 use mz_ore::cast::CastFrom;
-use mz_ore::num::Overflowing;
 use mz_ore::str::StrExt;
+use mz_ore::Overflowing;
 use mz_proto::{IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
 use ordered_float::OrderedFloat;
 use proptest::prelude::*;
